Change helpers structure
